Updating helpers and adding esxi block
