+++ powerflex/provider/sdc_host_resource.go
@@ -2,10 +2,6 @@ package provider
 
 import (
 	"context"
-	"fmt"
-	"regexp"
-	"strings"
-	"terraform-provider-powerflex/client"
 	"terraform-provider-powerflex/powerflex/helper"
 	"terraform-provider-powerflex/powerflex/models"
 
@@ -13,7 +9,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework-validators/listvalidator"
 	"github.com/hashicorp/terraform-plugin-framework-validators/resourcevalidator"
 	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
-	"github.com/hashicorp/terraform-plugin-framework/diag"
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
@@ -21,7 +16,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/types"
-	"github.com/hashicorp/terraform-plugin-framework/types/basetypes"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 )
 
@@ -29,20 +23,7 @@ var (
 	_ resource.Resource              = &sdcHostResource{}
 	_ resource.ResourceWithConfigure = &sdcHostResource{}
 	// _ resource.ResourceWithImportState = &sdcHostResource{}
-	_ client.Logger = &provisionerLogger{}
 )
-
-type provisionerLogger struct {
-	ctx context.Context
-}
-
-func (l *provisionerLogger) Printf(format string, v ...any) {
-	tflog.Info(l.ctx, fmt.Sprintf(format, v...))
-}
-
-func (l *provisionerLogger) Println(v ...any) {
-	tflog.Info(l.ctx, fmt.Sprint(v...))
-}
 
 // NewSDCResource is a helper function to simplify the provider implementation.
 func NewSDCHostResource() resource.Resource {
@@ -71,6 +52,28 @@ func (r *sdcHostResource) ConfigValidators(ctx context.Context) []resource.Confi
 		// 	remotePath.AtName("password"),
 		// 	remotePath.AtName("ca_cert"),
 		// ),
+	}
+}
+
+func (r *sdcHostResource) ValidateConfig(ctx context.Context, req resource.ValidateConfigRequest, resp *resource.ValidateConfigResponse) {
+	// Retrieve values from plan
+	var cfg models.SdcHostModel
+
+	diags := req.Config.Get(ctx, &cfg)
+
+	resp.Diagnostics.Append(diags...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	// for esxi, make sure that esxi block is provided (not null)
+	// Note: cfg.OS cannot be null (required field)
+	if !cfg.OS.IsUnknown() && cfg.OS.ValueString() == "esxi" && cfg.Esxi.IsNull() {
+		resp.Diagnostics.AddAttributeError(
+			path.Root("esxi"),
+			"Esxi block is required for esxi SDC",
+			"",
+		)
 	}
 }
 
@@ -147,19 +150,38 @@ func (r *sdcHostResource) Schema(_ context.Context, _ resource.SchemaRequest, re
 						MarkdownDescription: "Remote Login private key of the SDC server.",
 						Optional:            true,
 					},
+					"dir": schema.StringAttribute{
+						Description: "Directory on the SDC server to upload packages to." +
+							" Defaults to `/tmp` on Unix.",
+						MarkdownDescription: "Directory on the SDC server to upload packages to." +
+							" Defaults to `/tmp` on Unix.",
+						Optional: true,
+					},
 				},
 			},
-			"package_base64": schema.StringAttribute{
-				Description:         "Package to be installed on the SDC in its base64 format.",
-				MarkdownDescription: "Package to be installed on the SDC in its base64 format.",
+			"esxi": schema.SingleNestedAttribute{
+				Description:         "Details of the SDC host if the `os_family` is `esxi`.",
+				MarkdownDescription: "Details of the SDC host if the `os_family` is `esxi`.",
+				Optional:            true,
+				Attributes: map[string]schema.Attribute{
+					"guid": schema.StringAttribute{
+						Description:         "GUID of the SDC.",
+						MarkdownDescription: "GUID of the SDC.",
+						Required:            true,
+					},
+				},
+			},
+			"package_path": schema.StringAttribute{
+				Description:         "Full path (on local machine) of the package to be installed on the SDC.",
+				MarkdownDescription: "Full path (on local machine) of the package to be installed on the SDC.",
 				Required:            true,
 				Validators: []validator.String{
 					stringvalidator.LengthAtLeast(1),
 				},
 			},
-			"drv_cfg_base64": schema.StringAttribute{
-				Description:         "Driver Configuration file for the SDC in its base64 format.",
-				MarkdownDescription: "Driver Configuration file for the SDC in its base64 format.",
+			"drv_cfg_path": schema.StringAttribute{
+				Description:         "Full path (on local machine) of the driver Configuration file for the SDC.",
+				MarkdownDescription: "Full path (on local machine) of the driver Configuration file for the SDC.",
 				Required:            true,
 				Validators: []validator.String{
 					stringvalidator.LengthAtLeast(1),
@@ -208,149 +230,6 @@ func (r *sdcHostResource) Configure(_ context.Context, req resource.ConfigureReq
 	r.system = system
 }
 
-func (r *sdcHostResource) GetSshProvisioner(ctx context.Context, plan models.SdcHostModel) (*client.SshProvisioner, error) {
-	var remote models.SdcHostRemoteModel
-	plan.Remote.As(ctx, &remote, basetypes.ObjectAsOptions{UnhandledNullAsEmpty: true, UnhandledUnknownAsEmpty: true})
-	return client.NewSshProvisioner(client.SshProvisionerConfig{
-		IP:         plan.Host.ValueString(),
-		Username:   remote.User,
-		Password:   remote.Password,
-		PrivateKey: remote.PrivateKey,
-	}, &provisionerLogger{ctx: ctx})
-}
-
-func (r *sdcHostResource) GetMdmIps(ctx context.Context, plan models.SdcHostModel) []string {
-	var mdmIps []string
-	plan.MdmIPs.ElementsAs(ctx, &mdmIps, true)
-	return mdmIps
-}
-
-// createEsxi creates an esxi SDC host
-func (r *sdcHostResource) createEsxi(ctx context.Context, plan models.SdcHostModel) diag.Diagnostics {
-	var respDiagnostics diag.Diagnostics
-	sshP, err := r.GetSshProvisioner(ctx, plan)
-	if err != nil {
-		respDiagnostics.AddError(
-			"Error connecting to host",
-			err.Error(),
-		)
-		return respDiagnostics
-	}
-	defer sshP.Close()
-
-	// upload sw
-	scpProv := client.NewScpProvisioner(sshP)
-	err = scpProv.Upload(plan.Pkg.ValueString(), "/tmp/package.zip", "")
-	if err != nil {
-		respDiagnostics.AddError(
-			"Error uploading package",
-			err.Error(),
-		)
-		return respDiagnostics
-	}
-
-	// install sw
-	esxi := client.NewEsxCli(sshP)
-	pkgInstallCmd := client.VibInstallCommand{
-		ZipFile:  "/tmp/package.zip",
-		SigCheck: true,
-	}
-	op, err := esxi.SoftwareInstall(pkgInstallCmd)
-	if err != nil {
-		respDiagnostics.AddError(
-			"Error installing package",
-			err.Error()+"\n"+op,
-		)
-		return respDiagnostics
-	}
-
-	// reboot
-	err = sshP.RebootUnix()
-	if err != nil {
-		respDiagnostics.AddError(
-			"Error rebooting",
-			err.Error(),
-		)
-		return respDiagnostics
-	}
-
-	// check sw
-	tflog.Info(ctx, "Checking for installed sdc package")
-	sdc, err := esxi.GetSoftwareByNameRegex(regexp.MustCompile(".*sdc.*"))
-	if err != nil {
-		respDiagnostics.AddError(
-			"Error checking for installed sdc package",
-			err.Error(),
-		)
-		return respDiagnostics
-	}
-	tflog.Info(ctx, fmt.Sprintf("Installed SDC package is %s", sdc))
-
-	tflog.Info(ctx, "Setting scini module parameters")
-	params := map[string]string{
-		"IoctlIniGuidStr":        "87254810-e3cc-4c0b-87a7-91d3476b9ec7",
-		"IoctlMdmIPStr":          strings.Join(r.GetMdmIps(ctx, plan), ","),
-		"bBlkDevIsPdlActive":     "1",
-		"blkDevPdlTimeoutMillis": "60000",
-	}
-	if op, err := esxi.SetModuleParameters("scini", params); err != nil {
-		respDiagnostics.AddError(
-			"Error setting module parameters",
-			err.Error()+"\n"+op,
-		)
-		return respDiagnostics
-	}
-	tflog.Info(ctx, "Scini module parameters set")
-
-	// reboot
-	err = sshP.RebootUnix()
-	if err != nil {
-		respDiagnostics.AddError(
-			"Error rebooting",
-			err.Error(),
-		)
-		return respDiagnostics
-	}
-
-	// load esxi kernel modules
-	tflog.Info(ctx, "Loading vmk modules")
-	op, err = sshP.Run("vmkload_mod -l")
-	if err != nil {
-		respDiagnostics.AddError(
-			"Error loading vmk modules",
-			err.Error(),
-		)
-		return respDiagnostics
-	}
-	tflog.Info(ctx, "Finished loading vmk modules")
-	tflog.Debug(ctx, op)
-
-	// upload driver config
-	// recreate scpProvisioner
-	scpProv = client.NewScpProvisioner(sshP)
-	tflog.Info(ctx, "Uploading driver config")
-	err = scpProv.Upload(plan.DrvCfg.ValueString(), "/tmp/drv_cfg", "0755")
-	if err != nil {
-		respDiagnostics.AddError(
-			"Error uploading package",
-			err.Error(),
-		)
-		return respDiagnostics
-	}
-	// query mdms via drv cfg
-	tflog.Info(ctx, "Querying mdm ips via drv cfg")
-	op, err = sshP.Run("/tmp/drv_cfg --query_mdm")
-	if err != nil {
-		respDiagnostics.AddError(
-			"Error querying mdm ips via drv cfg",
-			err.Error()+"\n"+op,
-		)
-		return respDiagnostics
-	}
-
-	return respDiagnostics
-}
-
 // Create creates the resource and sets the initial Terraform state.
 func (r *sdcHostResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
 	// Retrieve values from plan
@@ -363,9 +242,13 @@ func (r *sdcHostResource) Create(ctx context.Context, req resource.CreateRequest
 		return
 	}
 
+	resHelper := helper.SdcHostResource{
+		System: r.system,
+	}
+
 	// install software
 	if plan.OS.ValueString() == "esxi" {
-		resp.Diagnostics.Append(r.createEsxi(ctx, plan)...)
+		resp.Diagnostics.Append(resHelper.CreateEsxi(ctx, plan)...)
 	}
 
 	if resp.Diagnostics.HasError() {
@@ -373,7 +256,7 @@ func (r *sdcHostResource) Create(ctx context.Context, req resource.CreateRequest
 	}
 
 	// read unconfigured SDC state after installation
-	currState, err := r.readSDCHost(ctx, plan)
+	currState, err := resHelper.ReadSDCHost(ctx, plan)
 	if err != nil {
 		resp.Diagnostics.AddError(
 			"Error reading SDC state",
@@ -388,7 +271,7 @@ func (r *sdcHostResource) Create(ctx context.Context, req resource.CreateRequest
 	}
 
 	// configure SDC via API
-	err = r.setSDCParams(ctx, plan, currState)
+	err = resHelper.SetSDCParams(ctx, plan, currState)
 	if err != nil {
 		resp.Diagnostics.AddError(
 			"Error setting SDC parameters",
@@ -398,7 +281,7 @@ func (r *sdcHostResource) Create(ctx context.Context, req resource.CreateRequest
 	}
 
 	// read final state of SDC and set state
-	state, err := r.readSDCHost(ctx, plan)
+	state, err := resHelper.ReadSDCHost(ctx, plan)
 	if err != nil {
 		resp.Diagnostics.AddError(
 			"Error reading SDC state",
@@ -411,42 +294,6 @@ func (r *sdcHostResource) Create(ctx context.Context, req resource.CreateRequest
 	resp.Diagnostics.Append(diags...)
 }
 
-func (r *sdcHostResource) setSDCParams(ctx context.Context, plan, state models.SdcHostModel) error {
-	// set name
-	if plan.Name.ValueString() != state.Name.ValueString() && !plan.Name.IsUnknown() {
-		tflog.Info(ctx, "Setting SDC name")
-		if _, err := r.system.ChangeSdcName(state.ID.ValueString(), plan.Name.ValueString()); err != nil {
-			return fmt.Errorf("error setting SDC name: %w", err)
-		}
-		tflog.Info(ctx, "SDC name set")
-	}
-
-	// set Performance Profile
-	if plan.PerformanceProfile.ValueString() != state.PerformanceProfile.ValueString() && !plan.PerformanceProfile.IsUnknown() {
-		tflog.Info(ctx, "Setting SDC performance profile")
-		if _, err := r.system.ChangeSdcPerfProfile(state.ID.ValueString(), plan.PerformanceProfile.ValueString()); err != nil {
-			return fmt.Errorf("error setting SDC performance profile: %w", err)
-		}
-		tflog.Info(ctx, "SDC performance profile set")
-	}
-
-	return nil
-}
-
-func (r *sdcHostResource) readSDCHost(ctx context.Context, state models.SdcHostModel) (models.SdcHostModel, error) {
-	// get SDC by IP
-	tflog.Info(ctx, "Finding SDC by IP")
-	sdcData, err := r.system.FindSdc("SdcIP", state.Host.ValueString())
-	if err != nil {
-		return state, fmt.Errorf("error finding SDC by IP %s: %w", state.Host.ValueString(), err)
-	}
-	tflog.Info(ctx, "Found SDC by IP")
-	state.ID = types.StringValue(sdcData.Sdc.ID)
-	state.PerformanceProfile = types.StringValue(sdcData.Sdc.PerfProfile)
-	state.Name = types.StringValue(sdcData.Sdc.Name)
-	return state, nil
-}
-
 // Read refreshes the Terraform state with the latest data.
 func (r *sdcHostResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
 	// Retrieve values from state
@@ -457,7 +304,11 @@ func (r *sdcHostResource) Read(ctx context.Context, req resource.ReadRequest, re
 		return
 	}
 
-	newState, err := r.readSDCHost(ctx, state)
+	resHelper := helper.SdcHostResource{
+		System: r.system,
+	}
+
+	newState, err := resHelper.ReadSDCHost(ctx, state)
 	if err != nil {
 		resp.Diagnostics.AddError(
 			"Error refreshing SDC state",
@@ -510,8 +361,12 @@ func (r *sdcHostResource) Update(ctx context.Context, req resource.UpdateRequest
 		return
 	}
 
+	resHelper := helper.SdcHostResource{
+		System: r.system,
+	}
+
 	// configure SDC via API
-	err := r.setSDCParams(ctx, plan, currState)
+	err := resHelper.SetSDCParams(ctx, plan, currState)
 	if err != nil {
 		resp.Diagnostics.AddError(
 			"Error setting SDC parameters",
@@ -521,7 +376,7 @@ func (r *sdcHostResource) Update(ctx context.Context, req resource.UpdateRequest
 	}
 
 	// read final state of SDC and set state
-	state, err := r.readSDCHost(ctx, plan)
+	state, err := resHelper.ReadSDCHost(ctx, plan)
 	if err != nil {
 		resp.Diagnostics.AddError(
 			"Error reading SDC state",
@@ -534,54 +389,6 @@ func (r *sdcHostResource) Update(ctx context.Context, req resource.UpdateRequest
 	resp.Diagnostics.Append(diags...)
 }
 
-func (r *sdcHostResource) deleteEsxi(ctx context.Context, state models.SdcHostModel) diag.Diagnostics {
-	var respDiagnostics diag.Diagnostics
-	// Disconnect from PowerFlex
-	tflog.Info(ctx, "Logging into host...")
-	sshP, err := r.GetSshProvisioner(ctx, state)
-	if err != nil {
-		respDiagnostics.AddError(
-			"Error connecting to host",
-			err.Error(),
-		)
-		return respDiagnostics
-	}
-	defer sshP.Close()
-
-	tflog.Info(ctx, "Checking for installed sdc package")
-	esxi := client.NewEsxCli(sshP)
-	sdc, err := esxi.GetSoftwareByNameRegex(regexp.MustCompile(".*sdc.*"))
-	if err != nil {
-		respDiagnostics.AddError(
-			"Error checking for installed sdc package",
-			err.Error(),
-		)
-		return respDiagnostics
-	}
-	tflog.Info(ctx, fmt.Sprintf("Installed SDC package is %v+", sdc))
-
-	op, err := esxi.SoftwareRmv(sdc.Name)
-	if err != nil {
-		respDiagnostics.AddError(
-			"Error removing sdc package",
-			err.Error()+"\n"+op,
-		)
-		return respDiagnostics
-	}
-	tflog.Info(ctx, fmt.Sprintf("sdc package removed: %s", op))
-
-	err = sshP.RebootUnix()
-	if err != nil {
-		respDiagnostics.AddError(
-			"Error rebooting host",
-			err.Error(),
-		)
-		return respDiagnostics
-	}
-
-	return respDiagnostics
-}
-
 // Delete deletes the resource and removes the Terraform state on success.
 func (r *sdcHostResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
 	// Retrieve values from state
@@ -592,9 +399,13 @@ func (r *sdcHostResource) Delete(ctx context.Context, req resource.DeleteRequest
 		return
 	}
 
+	resHelper := helper.SdcHostResource{
+		System: r.system,
+	}
+
 	// remove software
 	if state.OS.ValueString() == "esxi" {
-		resp.Diagnostics.Append(r.deleteEsxi(ctx, state)...)
+		resp.Diagnostics.Append(resHelper.DeleteEsxi(ctx, state)...)
 	}
 
 	if resp.Diagnostics.HasError() {
